feat: endpoint funcionario/medico/iniciar/consulta | fix: tipagem, modelagem da coluna

+++ src/main/java/com/healthapi/health_api/domain/Doctor.java
@@ -12,7 +12,7 @@ public class Doctor extends Worker implements Serializable {
     private String crm;
     private String specialty;
 
-    public Doctor(Long id, String cpf, String name, Date birthDate, String phoneNumber, String cep, String cidade, String uf, String logradouro, String numLogradouro, String complement, String email, Long workerId, String login, String password, String workerStatus, Date workerStartDate, String crm, String specialty) {
+    public Doctor(int id, String cpf, String name, Date birthDate, String phoneNumber, String cep, String cidade, String uf, String logradouro, String numLogradouro, String complement, String email, int workerId, String login, String password, String workerStatus, Date workerStartDate, String crm, String specialty) {
         super(id, cpf, name, birthDate, phoneNumber, cep, cidade, uf, logradouro, numLogradouro, complement, email, workerId, login, password, workerStatus, workerStartDate);
         this.crm = crm;
         this.specialty = specialty;
