+++ src/main/java/com/healthapi/health_api/dto/LoginResponseDTO.java
@@ -2,7 +2,7 @@ package com.healthapi.health_api.dto;
 
 public record LoginResponseDTO(
 
-    Long workerId,
+    int workerId,
 
     String login,
