+++ src/main/java/com/healthapi/health_api/dto/RegistrarFuncResponseDTO.java
@@ -2,7 +2,7 @@ package com.healthapi.health_api.dto;
 
 public record RegistrarFuncResponseDTO(
 
-    Long workerId,
+    int workerId,
 
     String login,
